Rework ptracer
The previous version failed miserably on threaded applications. This
version collects all threads so we can handle the case where a non
tracing thread trips over a software breakpoint used for stepping over
larx/stcx.

We now use PTRACE_SEIZE and follow fork/clone etc.

This version also uses perf events to fast forward, which is much
faster than single stepping. As a bonus, it uses the fast forwarding
code to determine which thread in a process is active, so you don't
end up trying to trace an idle thread.