qtbuild: qtrace builder tool
qtbuild can be used to edit a .qt. This is done by turning the .qt into
a special qtbuild type .S file, assembling that to a .o, and then using
qtbuild to turn that back into a .qt file.

This is done by adding annotations to the .S files that creates data
in elf sections that specify trace instruction and data addresses that
correspond with the assembled instructions.

It's a bit clunky but pretty simple and mostly uses existing tools.

This uses branch and ppc_storage from ptracer, which should eventually
be librified. Some consolidation and changes are made there.

This also adds a qtbuild output mode for qtrace2dis so that you can
create a qtbuild .S file from a .qt that can be edited and then built
back into a .qt.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>