htmdecoder: Retry when we encounter errors or record
Currently if we encounter an error on instruction metadata (ie.
instruction/data real/effective address), we say it's an invalid
record and move on.

This changes the policy to not mark this as an error. We now rewind to
try the metadata as the first record in a block. This seems to
sometimes occur in traces.

Signed-off-by: Michael Neuling <mikey@neuling.org>