htmdecoder: Integrate new software TLB
When a core trace is generated on a POWER8, it only spits out the real
address for branches when the address misses in the ERAT. This means
that most branches don't get a real address reported.

Unfortunately most qtrace tools expect the instruction real address
for every branch.

This adds a TLB emulation library that can be used to cache the
translations as we get them from the trace and then uses them later as
needed.

The code is written conservatively with a bunch of asserts when it
detects any corruption in the TLB. Every time the TLB is changed, we
run a validation check.

It also only does a linear lookup in the TLB for translations.

Despite this, benchmarking has shown this adds negligibly overhead to
the processing of a trace.

Signed-off-by: Michael Neuling <mikey@neuling.org>