htmdecoder: Don't continue to process record if it's corrupted
Currently if we hit a corrupted record, we continue processing the
current record. This can lead to problems since it's known broken.

This cleans up things and stops processing the record data that may be
corrupted. This means we can continue processing core files that are
slightly corrupted (rather than asserting).

This is implemented in a fragile way for now, but until we refactor
this code this is a minimal fix.

Signed-off-by: Michael Neuling <mikey@neuling.org>