qtdis/htmdecoder: Add instruction mix option -i
This adds an option to qtdis and htmdecoder to print the instruction
mix. Output looks like this:

     Instruction mix:
                           ld          2628520  12.42%
                          std          1885721   8.91%
                         addi          1466604   6.93%
                           mr          1287942   6.08%
                         beq-           951806   4.50%
                          add           751481   3.55%
                          lwz           615261   2.91%
                           li           564146   2.66%
                       clrldi           563695   2.66%
                         bne-           522009   2.47%
                        cmpdi           479051   2.26%
                         lwzx           449124   2.12%

For the opcode decode, this pulls in from binutils-gdb
(SHA1:4a5ee99b2f) opcodes/ppc-opc.c include/opcode/ppc.h. Minor
modifications are made to make it compile.

It's been optimised to be fast by including a cache of opcode
decodes. Kudos to Amitay for this.

Signed-off-by: Michael Neuling <mikey@neuling.org>