qtdis: Fix segv
When added the instruction mix option in:
  b71a401b49 qtdis/htmdecoder: Add instruction mix option -i
we imported a bunch of code from binutils. This seems to be causing
segvs when we now use disasm() libraries. This seems to be due to a
namespace collision.

We need a better solution long term so it's easier to import newer
versions of this code, but for now rename a bunch of things so we
avoid this segv.

Signed-off-by: Michael Neuling <mikey@neuling.org>