Add the Exception Termination Code to qtrace library
Currently qtwriter outputs the termination code of a terminating record
as zero unless the previous instruction had a branch opcode. That means
if an exception is taken the termination code will be zero.

This causes issues with qtreader. If a record has
QTRACE_TERMINATION_PRESENT set and has a termination code of zero then
this is an error which will halt processing.

For example:

	$ qtdis -s example-with-exception.qt
	Inconsistent branch

	Instructions:
	  Total           23709
	    System             526	  2.22% of Total
	      Idle               0	  0.00% of System
	    OPAL                 0	  0.00% of Total
	    User             23183	 97.78% of Total
	      Bin            21019	 90.67% of User
	      Lib             2164	  9.33% of User

	Context Switches:        0

	Exceptions:              1
		System Call/HCALL	1

	Syscalls:                1
			    exit	1

	OPAL Calls:              0

	HCALLS Calls:            0

Fix this by adding the QTRACE_EXCEPTION termination code from the qtrace
format to qtwriter and qtreader. qtraces previously made with htmdecoder
should be regenerated to correctly terminate exceptions.

Signed-off-by: Jordan Niethe <jniethe5@gmail.com>