htm: Add a host real to guest real address map
In some unclear circumstances, a page walk will resume from the host
real address of a PDE/PTE, while the guest real address is unknown.

It seems to occur when the host page size is bigger and there has
already been a another guest real to host real translation in that
range.

To create the missing guest PTE/PTE, keep a mapping of host real
addresses to guest real addresses.

Signed-off-by: Jordan Niethe <jniethe5@gmail.com>